Compiler Warning Cleanup Use uint32_t for rs2_intrinsics width height. Remove unnecessary casting after this change. This ripples into changing for loops to use autos. (in case this type changes later on) Several changes to the occlusion_filter to account for unsigned pixel width and height. Might want uint32_t for these??? Changed bpp to unsigned int. Other code uses size_t which seems overkill. Ideally defined the same everywhere.

+++ src/ds5/ds5-private.cpp
@@ -169,8 +169,8 @@ namespace librealsense
 
             // Calculate specific intrinsic parameters based on the normalized intrinsic and the sensor's resolution
             rs2_intrinsics calc_intrinsic{
-                static_cast<int>(width),
-                static_cast<int>(height),
+                width,
+                height,
                 ((1 + intrin(2, 0))*width) / 2.f,
                 ((1 + intrin(2, 1))*height) / 2.f,
                 intrin(0, 0) * width / 2.f,
